Reorganize test contracts and rename CEP78 to Cep78
