+++ modules/src/cep78/mod.rs
@@ -12,3 +12,4 @@ mod settings;
 mod tests;
 pub mod token;
 mod whitelist;
+mod utils;
