+++ modules/src/cep78/tests/metadata.rs
@@ -13,13 +13,14 @@ use crate::cep78::{
     },
     tests::{
         utils::{
-            TestContractHostRef, MALFORMED_META_DATA, TEST_PRETTY_CEP78_METADATA,
+            MALFORMED_META_DATA, TEST_PRETTY_CEP78_METADATA,
             TEST_PRETTY_UPDATED_CEP78_METADATA
         },
         TEST_CUSTOM_METADATA, TEST_CUSTOM_METADATA_SCHEMA, TEST_CUSTOM_UPDATED_METADATA,
         TOKEN_HASH
     },
-    token::CEP78HostRef
+    token::Cep78HostRef,
+    utils::MockContractHostRef
 };
 
 use super::{
@@ -36,7 +37,7 @@ fn should_prevent_update_in_immutable_mode() {
         .metadata_mutability(MetadataMutability::Immutable)
         .ownership_mode(OwnershipMode::Transferable)
         .build();
-    let mut contract = CEP78HostRef::deploy(&env, args);
+    let mut contract = Cep78HostRef::deploy(&env, args);
     contract.mint(
         env.get_account(0),
         TEST_PRETTY_721_META_DATA.to_string(),
@@ -58,15 +59,15 @@ fn should_prevent_update_in_immutable_mode() {
 
 #[test]
 fn should_prevent_install_with_hash_identifier_in_mutable_mode() {
-    let env = odra_test::env();
-    let args = default_args_builder()
-        .nft_metadata_kind(NFTMetadataKind::NFT721)
-        .identifier_mode(NFTIdentifierMode::Hash)
-        .metadata_mutability(MetadataMutability::Mutable)
-        .build();
-    let _contract = CEP78HostRef::deploy(&env, args);
+    // let env = odra_test::env();
+    // let args = default_args_builder()
+    //     .nft_metadata_kind(NFTMetadataKind::NFT721)
+    //     .identifier_mode(NFTIdentifierMode::Hash)
+    //     .metadata_mutability(MetadataMutability::Mutable)
+    //     .build();
+    // let _contract = Cep78HostRef::deploy(&env, args);
     // Should be possible to verify errors at installation time
-    // assert_eq!(CEP78HostRef::deploy(&env, args), Err(CEP78Error::InvalidMetadataMutability));
+    // assert_eq!(Cep78HostRef::deploy(&env, args), Err(CEP78Error::InvalidMetadataMutability));
 }
 
 #[test]
@@ -78,7 +79,7 @@ fn should_prevent_update_for_invalid_metadata() {
         .metadata_mutability(MetadataMutability::Mutable)
         .ownership_mode(OwnershipMode::Transferable)
         .build();
-    let mut contract = CEP78HostRef::deploy(&env, args);
+    let mut contract = Cep78HostRef::deploy(&env, args);
     contract.mint(
         env.get_account(0),
         TEST_PRETTY_721_META_DATA.to_string(),
@@ -108,7 +109,7 @@ fn should_prevent_metadata_update_by_non_owner_key() {
         .metadata_mutability(MetadataMutability::Mutable)
         .ownership_mode(OwnershipMode::Transferable)
         .build();
-    let mut contract = CEP78HostRef::deploy(&env, args);
+    let mut contract = Cep78HostRef::deploy(&env, args);
     let token_owner = *contract.address();
     contract.mint(
         token_owner,
@@ -148,7 +149,7 @@ fn should_allow_update_for_valid_metadata_based_on_kind(
         .json_schema(json_schema)
         .events_mode(EventsMode::CES)
         .build();
-    let mut contract = CEP78HostRef::deploy(&env, args);
+    let mut contract = Cep78HostRef::deploy(&env, args);
     let token_owner = env.get_account(0);
 
     let custom_metadata = serde_json::to_string_pretty(&*TEST_CUSTOM_METADATA)
@@ -259,7 +260,7 @@ fn should_update_metadata_for_custom_validated_using_token_id() {
 #[test]
 fn should_get_metadata_using_token_id() {
     let env = odra_test::env();
-    let mut minting_contract = TestContractHostRef::deploy(&env, NoArgs);
+    let mut minting_contract = MockContractHostRef::deploy(&env, NoArgs);
     let contract_whitelist = vec![*minting_contract.address()];
     let args = default_args_builder()
         .holder_mode(NFTHolderMode::Contracts)
@@ -268,7 +269,7 @@ fn should_get_metadata_using_token_id() {
         .minting_mode(MintingMode::Acl)
         .acl_white_list(contract_whitelist)
         .build();
-    let mut contract = CEP78HostRef::deploy(&env, args);
+    let mut contract = Cep78HostRef::deploy(&env, args);
     minting_contract.set_address(contract.address());
     let token_id = 0u64;
 
@@ -286,7 +287,7 @@ fn should_get_metadata_using_token_id() {
 #[test]
 fn should_get_metadata_using_token_metadata_hash() {
     let env = odra_test::env();
-    let mut minting_contract = TestContractHostRef::deploy(&env, NoArgs);
+    let mut minting_contract = MockContractHostRef::deploy(&env, NoArgs);
     let contract_whitelist = vec![*minting_contract.address()];
     let args = default_args_builder()
         .identifier_mode(NFTIdentifierMode::Hash)
@@ -297,7 +298,7 @@ fn should_get_metadata_using_token_metadata_hash() {
         .minting_mode(MintingMode::Acl)
         .acl_white_list(contract_whitelist)
         .build();
-    let mut contract = CEP78HostRef::deploy(&env, args);
+    let mut contract = Cep78HostRef::deploy(&env, args);
     minting_contract.set_address(contract.address());
 
     assert!(
@@ -317,7 +318,7 @@ fn should_get_metadata_using_token_metadata_hash() {
 #[test]
 fn should_revert_minting_token_metadata_hash_twice() {
     let env = odra_test::env();
-    let mut minting_contract = TestContractHostRef::deploy(&env, NoArgs);
+    let mut minting_contract = MockContractHostRef::deploy(&env, NoArgs);
     let contract_whitelist = vec![*minting_contract.address()];
     let args = default_args_builder()
         .identifier_mode(NFTIdentifierMode::Hash)
@@ -328,7 +329,7 @@ fn should_revert_minting_token_metadata_hash_twice() {
         .minting_mode(MintingMode::Acl)
         .acl_white_list(contract_whitelist)
         .build();
-    let mut contract = CEP78HostRef::deploy(&env, args);
+    let mut contract = Cep78HostRef::deploy(&env, args);
     minting_contract.set_address(contract.address());
     assert!(
         contract.is_whitelisted(minting_contract.address()),
@@ -351,7 +352,7 @@ fn should_revert_minting_token_metadata_hash_twice() {
 #[test]
 fn should_get_metadata_using_custom_token_hash() {
     let env = odra_test::env();
-    let mut minting_contract = TestContractHostRef::deploy(&env, NoArgs);
+    let mut minting_contract = MockContractHostRef::deploy(&env, NoArgs);
     let contract_whitelist = vec![*minting_contract.address()];
     let args = default_args_builder()
         .identifier_mode(NFTIdentifierMode::Hash)
@@ -362,7 +363,7 @@ fn should_get_metadata_using_custom_token_hash() {
         .minting_mode(MintingMode::Acl)
         .acl_white_list(contract_whitelist)
         .build();
-    let mut contract = CEP78HostRef::deploy(&env, args);
+    let mut contract = Cep78HostRef::deploy(&env, args);
     minting_contract.set_address(contract.address());
 
     assert!(
@@ -382,7 +383,7 @@ fn should_get_metadata_using_custom_token_hash() {
 #[test]
 fn should_revert_minting_custom_token_hash_identifier_twice() {
     let env = odra_test::env();
-    let mut minting_contract = TestContractHostRef::deploy(&env, NoArgs);
+    let mut minting_contract = MockContractHostRef::deploy(&env, NoArgs);
     let contract_whitelist = vec![*minting_contract.address()];
     let args = default_args_builder()
         .identifier_mode(NFTIdentifierMode::Hash)
@@ -393,7 +394,7 @@ fn should_revert_minting_custom_token_hash_identifier_twice() {
         .minting_mode(MintingMode::Acl)
         .acl_white_list(contract_whitelist)
         .build();
-    let mut contract = CEP78HostRef::deploy(&env, args);
+    let mut contract = Cep78HostRef::deploy(&env, args);
     minting_contract.set_address(contract.address());
 
     assert!(
@@ -443,7 +444,7 @@ fn should_require_json_schema_when_kind_is_custom_validated() {
         .nft_metadata_kind(nft_metadata_kind)
         .json_schema("".to_string())
         .build();
-    let _contract = CEP78HostRef::deploy(&env, args);
+    let _contract = Cep78HostRef::deploy(&env, args);
 
     /*let error = builder.get_error().expect("must have error");
     support::assert_expected_error(error, 67, "json_schema is required")*/
@@ -458,7 +459,7 @@ fn should_not_require_json_schema_when_kind_is(nft_metadata_kind: NFTMetadataKin
         .nft_metadata_kind(nft_metadata_kind.clone())
         .json_schema("".to_string())
         .build();
-    let mut contract = CEP78HostRef::deploy(&env, args);
+    let mut contract = Cep78HostRef::deploy(&env, args);
 
     let original_metadata = match &nft_metadata_kind {
         NFTMetadataKind::CEP78 => TEST_PRETTY_CEP78_METADATA,
