+++ modules/src/cep78/tests/transfer.rs
@@ -15,12 +15,12 @@ use crate::cep78::{
     },
     tests::{
         default_args_builder,
-        utils::{TestContractHostRef, TEST_PRETTY_721_META_DATA}
+        utils::TEST_PRETTY_721_META_DATA
     },
-    token::CEP78HostRef
+    token::Cep78HostRef, utils::{MockContractHostRef, MockTransferFilterContractHostRef}
 };
 
-use super::utils::{self, TransferFilterContractHostRef};
+use super::utils;
 
 #[test]
 fn should_disallow_transfer_with_minter_or_assigned_ownership_mode() {
@@ -30,7 +30,7 @@ fn should_disallow_transfer_with_minter_or_assigned_ownership_mode() {
         .minting_mode(MintingMode::Installer)
         .owner_reverse_lookup_mode(OwnerReverseLookupMode::Complete)
         .build();
-    let mut contract = CEP78HostRef::deploy(&env, args);
+    let mut contract = Cep78HostRef::deploy(&env, args);
 
     let token_owner = env.get_account(0);
     contract.register_owner(Maybe::Some(token_owner));
@@ -67,7 +67,7 @@ fn should_transfer_token_from_sender_to_receiver() {
         .owner_reverse_lookup_mode(OwnerReverseLookupMode::Complete)
         .events_mode(EventsMode::CES)
         .build();
-    let mut contract = CEP78HostRef::deploy(&env, args);
+    let mut contract = Cep78HostRef::deploy(&env, args);
 
     let token_owner = env.get_account(0);
     contract.register_owner(Maybe::Some(token_owner));
@@ -126,7 +126,7 @@ fn approve_token_for_transfer_should_add_entry_to_approved_dictionary(
         .owner_reverse_lookup_mode(OwnerReverseLookupMode::Complete)
         .events_mode(EventsMode::CES)
         .build();
-    let mut contract = CEP78HostRef::deploy(&env, args);
+    let mut contract = Cep78HostRef::deploy(&env, args);
 
     let token_owner = env.get_account(0);
     contract.register_owner(Maybe::Some(token_owner));
@@ -186,7 +186,7 @@ fn revoke_token_for_transfer_should_remove_entry_to_approved_dictionary(
         .owner_reverse_lookup_mode(OwnerReverseLookupMode::Complete)
         .events_mode(EventsMode::CES)
         .build();
-    let mut contract = CEP78HostRef::deploy(&env, args);
+    let mut contract = Cep78HostRef::deploy(&env, args);
 
     let token_owner = env.get_account(0);
     contract.register_owner(Maybe::Some(token_owner));
@@ -245,7 +245,7 @@ fn should_disallow_approving_when_ownership_mode_is_minter_or_assigned() {
         .owner_reverse_lookup_mode(OwnerReverseLookupMode::Complete)
         .events_mode(EventsMode::CES)
         .build();
-    let mut contract = CEP78HostRef::deploy(&env, args);
+    let mut contract = Cep78HostRef::deploy(&env, args);
     let token_owner = env.get_account(0);
     contract.register_owner(Maybe::Some(token_owner));
     contract.mint(
@@ -269,7 +269,7 @@ fn should_be_able_to_transfer_token(env: HostEnv, operator: Option<Address>) {
         .owner_reverse_lookup_mode(OwnerReverseLookupMode::Complete)
         .events_mode(EventsMode::CES)
         .build();
-    let mut contract = CEP78HostRef::deploy(&env, args);
+    let mut contract = Cep78HostRef::deploy(&env, args);
     let token_owner = env.get_account(0);
     contract.register_owner(Maybe::Some(token_owner));
     contract.mint(
@@ -328,7 +328,7 @@ fn should_disallow_same_approved_account_to_transfer_token_twice() {
         .owner_reverse_lookup_mode(OwnerReverseLookupMode::Complete)
         .events_mode(EventsMode::CES)
         .build();
-    let mut contract = CEP78HostRef::deploy(&env, args);
+    let mut contract = Cep78HostRef::deploy(&env, args);
 
     let token_owner = env.get_account(0);
     contract.register_owner(Maybe::Some(token_owner));
@@ -384,7 +384,7 @@ fn should_disallow_to_transfer_token_using_revoked_hash(env: HostEnv, operator: 
         .ownership_mode(OwnershipMode::Transferable)
         .owner_reverse_lookup_mode(OwnerReverseLookupMode::Complete)
         .build();
-    let mut contract = CEP78HostRef::deploy(&env, args);
+    let mut contract = Cep78HostRef::deploy(&env, args);
 
     let token_owner = env.get_account(0);
     contract.register_owner(Maybe::Some(token_owner));
@@ -459,7 +459,7 @@ fn should_be_able_to_approve_with_deprecated_operator_argument() {}
 #[test]
 fn should_transfer_between_contract_to_account() {
     let env = odra_test::env();
-    let mut minting_contract = TestContractHostRef::deploy(&env, NoArgs);
+    let mut minting_contract = MockContractHostRef::deploy(&env, NoArgs);
     let contract_whitelist = vec![*minting_contract.address()];
     let args = default_args_builder()
         .owner_reverse_lookup_mode(OwnerReverseLookupMode::Complete)
@@ -470,7 +470,7 @@ fn should_transfer_between_contract_to_account() {
         .minting_mode(MintingMode::Acl)
         .acl_white_list(contract_whitelist)
         .build();
-    let mut contract = CEP78HostRef::deploy(&env, args);
+    let mut contract = Cep78HostRef::deploy(&env, args);
     minting_contract.set_address(contract.address());
 
     assert!(
@@ -502,7 +502,7 @@ fn should_prevent_transfer_when_caller_is_not_owner() {
         .ownership_mode(OwnershipMode::Transferable)
         .holder_mode(NFTHolderMode::Accounts)
         .build();
-    let mut contract = CEP78HostRef::deploy(&env, args);
+    let mut contract = Cep78HostRef::deploy(&env, args);
 
     let token_owner = env.get_account(0);
     let unauthorized_user = env.get_account(10);
@@ -537,7 +537,7 @@ fn should_transfer_token_in_hash_identifier_mode() {
         .ownership_mode(OwnershipMode::Transferable)
         .metadata_mutability(MetadataMutability::Immutable)
         .build();
-    let mut contract = CEP78HostRef::deploy(&env, args);
+    let mut contract = Cep78HostRef::deploy(&env, args);
 
     let token_owner = env.get_account(0);
     let new_owner = env.get_account(1);
@@ -559,13 +559,13 @@ fn should_transfer_token_in_hash_identifier_mode() {
 #[test]
 fn should_not_allow_non_approved_contract_to_transfer() {
     let env = odra_test::env();
-    let mut minting_contract = TestContractHostRef::deploy(&env, NoArgs);
+    let mut minting_contract = MockContractHostRef::deploy(&env, NoArgs);
     let args = default_args_builder()
         .owner_reverse_lookup_mode(OwnerReverseLookupMode::Complete)
         .holder_mode(NFTHolderMode::Mixed)
         .ownership_mode(OwnershipMode::Transferable)
         .build();
-    let mut contract = CEP78HostRef::deploy(&env, args);
+    let mut contract = Cep78HostRef::deploy(&env, args);
     minting_contract.set_address(contract.address());
     let token_owner = env.get_account(0);
     contract.register_owner(Maybe::Some(token_owner));
@@ -602,7 +602,7 @@ fn transfer_should_correctly_track_page_table_entries() {
         .nft_metadata_kind(NFTMetadataKind::Raw)
         .ownership_mode(OwnershipMode::Transferable)
         .build();
-    let mut contract = CEP78HostRef::deploy(&env, args);
+    let mut contract = Cep78HostRef::deploy(&env, args);
     let token_owner = env.get_account(0);
     let new_owner = env.get_account(1);
 
@@ -635,7 +635,7 @@ fn should_prevent_transfer_to_unregistered_owner() {
         .identifier_mode(NFTIdentifierMode::Ordinal)
         .metadata_mutability(MetadataMutability::Immutable)
         .build();
-    let mut contract = CEP78HostRef::deploy(&env, args);
+    let mut contract = Cep78HostRef::deploy(&env, args);
     let token_owner = env.get_account(0);
     contract.register_owner(Maybe::Some(token_owner));
     contract.mint(token_owner, "".to_string(), Maybe::None);
@@ -661,7 +661,7 @@ fn should_transfer_token_from_sender_to_receiver_with_transfer_only_reporting() 
         .owner_reverse_lookup_mode(OwnerReverseLookupMode::TransfersOnly)
         .ownership_mode(OwnershipMode::Transferable)
         .build();
-    let mut contract = CEP78HostRef::deploy(&env, args);
+    let mut contract = Cep78HostRef::deploy(&env, args);
     let token_owner = env.get_account(0);
     contract.register_owner(Maybe::Some(token_owner));
     env.set_caller(token_owner);
@@ -706,7 +706,7 @@ fn disallow_owner_to_approve_itself() {
     let args = default_args_builder()
         .ownership_mode(OwnershipMode::Transferable)
         .build();
-    let mut contract = CEP78HostRef::deploy(&env, args);
+    let mut contract = Cep78HostRef::deploy(&env, args);
     let token_owner = env.get_account(0);
     contract.register_owner(Maybe::Some(token_owner));
     contract.mint(
@@ -727,7 +727,7 @@ fn disallow_operator_to_approve_itself() {
     let args = default_args_builder()
         .ownership_mode(OwnershipMode::Transferable)
         .build();
-    let mut contract = CEP78HostRef::deploy(&env, args);
+    let mut contract = Cep78HostRef::deploy(&env, args);
     let token_owner = env.get_account(0);
     contract.register_owner(Maybe::Some(token_owner));
     contract.mint(
@@ -753,7 +753,7 @@ fn disallow_owner_to_approve_for_all_itself() {
     let args = default_args_builder()
         .ownership_mode(OwnershipMode::Transferable)
         .build();
-    let mut contract = CEP78HostRef::deploy(&env, args);
+    let mut contract = Cep78HostRef::deploy(&env, args);
     let token_owner = env.get_account(0);
     contract.register_owner(Maybe::Some(token_owner));
     contract.mint(
@@ -772,14 +772,13 @@ fn disallow_owner_to_approve_for_all_itself() {
 fn check_transfers_with_transfer_filter_contract_modes() {
     let env = odra_test::env();
 
-    let mut transfer_filter_contract: TransferFilterContractHostRef =
-        TransferFilterContractHostRef::deploy(&env, NoArgs);
+    let mut transfer_filter_contract = MockTransferFilterContractHostRef::deploy(&env, NoArgs);
     transfer_filter_contract.set_return_value(TransferFilterContractResult::DenyTransfer as u8);
     let args = default_args_builder()
         .ownership_mode(OwnershipMode::Transferable)
         .transfer_filter_contract_contract_key(*transfer_filter_contract.address())
         .build();
-    let mut contract = CEP78HostRef::deploy(&env, args);
+    let mut contract = Cep78HostRef::deploy(&env, args);
     let token_owner = env.get_account(0);
     contract.register_owner(Maybe::Some(token_owner));
     contract.mint(
@@ -846,12 +845,12 @@ fn check_transfers_with_transfer_filter_contract_modes() {
 #[test]
 fn should_disallow_transfer_from_contract_with_package_operator_mode_without_operator() {
     let env = odra_test::env();
-    let mut minting_contract = TestContractHostRef::deploy(&env, NoArgs);
+    let mut minting_contract = MockContractHostRef::deploy(&env, NoArgs);
     let args = default_args_builder()
         .ownership_mode(OwnershipMode::Transferable)
         .holder_mode(NFTHolderMode::Mixed)
         .build();
-    let mut contract = CEP78HostRef::deploy(&env, args);
+    let mut contract = Cep78HostRef::deploy(&env, args);
     minting_contract.set_address(contract.address());
     let token_owner = env.get_account(0);
     contract.register_owner(Maybe::Some(token_owner));
@@ -879,12 +878,12 @@ fn should_disallow_transfer_from_contract_without_package_operator_mode_with_pac
 #[test]
 fn should_allow_transfer_from_contract_with_package_operator_mode_with_operator() {
     let env = odra_test::env();
-    let mut minting_contract = TestContractHostRef::deploy(&env, NoArgs);
+    let mut minting_contract = MockContractHostRef::deploy(&env, NoArgs);
     let args = default_args_builder()
         .ownership_mode(OwnershipMode::Transferable)
         .holder_mode(NFTHolderMode::Mixed)
         .build();
-    let mut contract = CEP78HostRef::deploy(&env, args);
+    let mut contract = Cep78HostRef::deploy(&env, args);
     minting_contract.set_address(contract.address());
     let token_owner = env.get_account(0);
     contract.register_owner(Maybe::Some(token_owner));
@@ -914,12 +913,12 @@ fn should_disallow_package_operator_to_approve_without_package_operator_mode() {
 #[test]
 fn should_allow_package_operator_to_approve_with_package_operator_mode() {
     let env = odra_test::env();
-    let mut minting_contract = TestContractHostRef::deploy(&env, NoArgs);
+    let mut minting_contract = MockContractHostRef::deploy(&env, NoArgs);
     let args = default_args_builder()
         .ownership_mode(OwnershipMode::Transferable)
         .holder_mode(NFTHolderMode::Mixed)
         .build();
-    let mut contract = CEP78HostRef::deploy(&env, args);
+    let mut contract = Cep78HostRef::deploy(&env, args);
     minting_contract.set_address(contract.address());
     let token_owner = env.get_account(0);
     contract.register_owner(Maybe::Some(token_owner));
@@ -952,12 +951,12 @@ fn should_allow_package_operator_to_approve_with_package_operator_mode() {
 #[test]
 fn should_allow_account_to_approve_spender_with_package_operator() {
     let env = odra_test::env();
-    let minting_contract = TestContractHostRef::deploy(&env, NoArgs);
+    let minting_contract = MockContractHostRef::deploy(&env, NoArgs);
     let args = default_args_builder()
         .ownership_mode(OwnershipMode::Transferable)
         .holder_mode(NFTHolderMode::Mixed)
         .build();
-    let mut contract = CEP78HostRef::deploy(&env, args);
+    let mut contract = Cep78HostRef::deploy(&env, args);
     let token_owner = env.get_account(0);
     contract.register_owner(Maybe::Some(token_owner));
     contract.mint(
@@ -989,12 +988,12 @@ fn should_allow_account_to_approve_spender_with_package_operator() {
 #[test]
 fn should_allow_package_operator_to_revoke_with_package_operator_mode() {
     let env = odra_test::env();
-    let mut minting_contract = TestContractHostRef::deploy(&env, NoArgs);
+    let mut minting_contract = MockContractHostRef::deploy(&env, NoArgs);
     let args = default_args_builder()
         .ownership_mode(OwnershipMode::Transferable)
         .holder_mode(NFTHolderMode::Mixed)
         .build();
-    let mut contract = CEP78HostRef::deploy(&env, args);
+    let mut contract = Cep78HostRef::deploy(&env, args);
     minting_contract.set_address(contract.address());
     let token_owner = env.get_account(0);
     contract.register_owner(Maybe::Some(token_owner));
